Use pthread_atfork to reinstall the heartbeat thread post fork.
David Lewis suggested:
On Wed, May 21, 2014 at 04:35:33PM +0200, Stephan Eggermont wrote: 
> PetitParser 1.6 instead of 1.5. 
> Tests only work when no Nautilus browser is open. Some notification problem when creating classes. 
> 
> OSProcess ci has lots of failing tests, so we don't know if it actually works??? 
> 
> Diego & Stephan 

The tests for OSProcess require VM support. Until recently, an interpreter VM was 
required to run the full OSProcess test suite because it relies heavily on #forkSqueak 
to generate cooperating VM processes for the tests. 

Eliot has recently added an enhancement to Cog that provides the required support. When 
that change is added to the Pharo VM, the tests will run. 

The changes are in SVN in branches/Cog/platforms/unix/vm/ and the commit notice is at: 

http://lists.squeakfoundation.org/pipermail/vm-dev/2014-March/014964.html