VMMaker.oscog-eem.1022:Cope with the funky primitives generated fromSmalltalk methods in computing accessor depths.
VMMaker.oscog-eem.1021:Fix in-image compilation for the at:put: primitive.Bow to the inevitable and add a manual override forprimitive accessor depth.  Use it in at: & at:put:.If manual override is not used then the automaticallycomputed depth is too deep because of contextaccess.

VMMaker.oscog-eem.1020:Fix monumental blunder in jitted store-check code.The saved registers mask needs to be the *or* ofthe value reg and the caller-saved registers, notthe *and*!!Merge with VMMaker.oscog-cb.1019's changes thatuse the method header flag bit to identify optimizedmethods.  rename HeaderFlagBitPosition andMethodHeaderOptimizedBitShift toMethodHeaderFlagBitPosition.Slang:Eliminate null expressions in and: and or: ifgenerateDeadCode is false.