VMMaker.oscog-eem.1060:And FloatMathPlugin needs floatExponentBits
VMMaker.oscog-eem.1059:Add support for <inline: #always> to mark thegenerated function as inline for use when Slangcan't inline (it currently doesn't inline multi-statement functions in conditionals).

VMMaker.oscog-eem.1058:Revisit primitiveTimesTwoPower.  Avoid complicationsin the 32-bit versions.  Clip the argument in the 64-bitversions and avoid the compiler warning with a cast.