VMMaker.oscog-eem.1068:...and add type inference support for at: rto rescueinlining in the Klatt plugin.Mark SpurSegmentManager>>initializeFromFreeChunks:as doNotGenerate to avoid a type inferrence error.
VMMaker.oscog-eem.1067:Back out of the wrong-headed attempt to fix thesymptoms of bad inlining rather than the cause byhacking comparison generation.Insist on signedness of actual expressions matchingthe signedness of formals to inline an expression,otherwise assign the expression to the inlined formal.Add support for type inferrence of simple arithmetic(+,-,*,/) but *not* >> << et al, so that inlining is stilldone for e.g. pointer arithmetic expressions.Refactor the inlining breakpoint machinery so one canbreak only at the point of inline as well as only at thepoint of testing for inlinability.Fix a bug in TMethod>>typeFor:in: which was causingthe failure to find the types of global variables.