VMMaker.oscog-eem.1096:Fix regression in generateObjectRepresentationTrampolines.Of /course/ ceScheduleScavengeTrampoline needs tobe a safe trampoline.
VMMaker.oscog-eem.1095:On ARM the ceTraceLinkedSend call needs toposh/pop the LinkReg around the call.

VMMaker.oscog-eem.1094:Make the ISA explicit in the VM generators, now thatwe're running the ARM simulator alongside the IA32one.  Make sure Cogit is initialized with optionswhen generating the interpreter.Fix assertClassOf:is: to handle immediates upstreamof isClassOfNonImm:equalTo:.Fix CogARMCompiler>>computeMaximumSize for Slang.