VMMaker.oscog-eem.1129:Add support for the VarBaseReg scheme to ARMMoveMbrR & MoveRMbr sicne these are used foraccessing the primTraceLogIndex.Add hasPCRelativeAddressing and for the ARM,isAddressRelativeToPC:.  We can use this to referto the method without generating a map entry.Mark the remaining relevant feature test methods inCogAbstractInstruciton as <inline: true>.
VMMaker.oscog-eem.1128:Oops!  remember to answer an abstract registerderived mask for ARM's callerSavedRegisterMask.Change the ARM reg assignments to use r0 forTempReg and r7 for ResultReceiverReg.At some stage (e.g. when Clément arrives to workon Sista) the register mask machinery may have tochange to use concrete register numbers to allowthe allocator to allocate more registers than canbe named as abstract registers.

VMMaker.oscog-eem.1127:Rename executeCogMethod:... to executeCogPIC:...Add support for a variable base register forreferring to the CoInterpreter variables frommachine code on RISCs (ARM).  Load the VarBaseRegin enilopmarts.  It must be a callee-saved reg.Add support for a 3 argument AndCqRR and use itin the entry-point sequence.  This has the nice side-effect of fixing the inability to differentiate thechecked and unchecked entry-points on ARM if weuse the MoveRR, AndCqR two instruction sequence ;)Nuke the unused and misleading callsAreRelative.Mark the other backEnd ability test methods as<inline: true>.Simulator:Add print cog methods with selector... to the utils menu.