Cog-eem.255:Include the offset in the relevant addressing modesin load and store exceptions in the ARM alien.
Cog-eem.254:Rewrite GdbARMAlien>>handleExecutionPrimitiveFailureIn:minimumAddress:to be simpler and faster, using the endCondition toidentify call, jump and return traps.No longer extract addresses from previous instructions(which presumes a code generation stategy), but accessthe underlying register state.  This is a RSISC after all!Rename registerStateNames to registerStateGetters toaccord with the IA32 Alien.

Cog-tpr.253:convert various methods to refer to 'sl' instead of 'r10' since that is what we print out in debugging messages. Clean up a couple of alien tests

Cog-eem.252:Fix comments, and add the return codes as class vars.