VMMaker.oscog-eem.1156:Integrate Nicolas' fixes that usepositiveMachineIntegerFor: for pointers.
VMMaker.oscog-eem.1155:and fix that compiler warning properly.  If one definesa function return type the arguments are typed there-in.

VMMaker.oscog-cb.1154:fix inline primitive for at: and at:put: to be 1 relative. This simplifies the JIT in 64 bits because the 1 relative correponds to the 1 word difference between baseHeader and first object field hence generating one less instruction.