VMMaker.oscog-tpr.1158:remove no longer needed stack usage tracing method. Replaced by improved decorateDisassembly:for:fromAddress:
VMMaker.oscog-eem.1157:Fix the num initial nils processing in StackToRegisterMappingCogitfor SistaV1.  The pushNilSize: functions need to take numInitialNilsas an additional argument so that those that have a single bytecodefor all nils can compute the size of the bytecode correctly.Nuke the unused initializeBytecodeTableForNewspeakV3PlusClosuresinitializer.  Fix in-image compilation by addingSpurMemoryManager class>>bytesPerOop