VMMaker.oscog-eem.1202:Fix regression in map machinery due to addingAnnotationExtension scheme.  findMapLocationForMcpc:inMethod:must not be confused by IsDisplacementX2N bytes.This is likely the cause of the recent crashes withr3308 and earlier.Lay the groundwork for 32-bit intra-zone jumps andcalls on ARM by introducing CallFull and JumpFull(and rewrites thereof) that are expected tospan the ful address space, leaving Call/JumpLongto span merely the 16mb code zone.  On x86CallFull and JumpFull simply default to Call/JumpLong.Tim will implement the ARM support soon.
VMMaker.oscog-cb.1201:Replace bytecode trapIfNotInstanceOf by jumpIfNotInstanceOfOrPop.Rewrote the JIT logic for traps to be able to write trap trampolines calls at the end of the cogMethod.

VMMaker.oscog-eem.1200:Refactor the slot store and store check machineryto take an inFrame: argument and hence deal withthe store check in genInnerPrimitiveAtPut: on ARM.Implement jumpLong[Conditional]ByteSize for ARMnoting that this is unsatisfactory.  We need togenerate "short" branches on ARM for PICs.