VMMaker.oscog-eem.1217:Fix printing of CogFooInstructions with code LastRTLCode.Make read-before-written initializer not initialize arrays.
VMMaker.oscog-cb.1216:fix a bug where some #== followed by a branch would not generate the compare instruction before the jumpZero.

VMMaker.oscog-cb.1215:Fixed machine code generation of #== in the sista cogit. simulated stack was imbalanced and I made a typo label -> Label

VMMaker.oscog-cb.1214:- Change the code generation of == for sistaCogit- change the coutner reg to be allocated instead of being a fixed abstract reg- fix a C code gen bug in code gen of #==- fix a bug in #== where the arg was not unforwarded