VMMaker.oscog-cb.1245:Simple refactoring for booleans moved to registers.
VMMaker.oscog-cb.1244:Changed #genStorePop: popBoolean LiteralVariable: litVarIndex so it uses register allocation instead of ReceiverResultReg and ClassReg.This method now does not require a frame if the assigned value is an unannotatableConstant. However, it still needs a frame in the other case due to the ceStoreCheck.

VMMaker.oscog-cb.1243:Removed a ssAllocateRequiredReg which was not needed.Basically refactored:| constVal |constVal := self ssTop maybeConstant.	(self ssTop type = SSConstant	 and: [(objectRepresentation shouldAnnotateObjectReference: constVal) not])	to:(objectRepresentation isUnannotatableConstant: self ssTop)so I can understand the methods and see if I can change them with better register allocation. I am interested especially in genStorePop: popBoolean LiteralVariable: litVarIndex which has the flag:	self flag: 'with better register allocation this wouldn''t need a frame.  e.g. use SendNumArgs instead of ReceiverResultReg'.

VMMaker.oscog-eem.1242:Fix the read-before-written initializer to not initializestructs (if the struct types can be identified).Make the allOldMarkedWeakObjectsOnWeaklingStackassert an eassert: (it /is/ extremely expensive) andcorrect some typos in eassert:.