VMMaker.oscog-eem.1303:Fix the regression with #== in the StackToRegisterMappingCogit;a spilled constant must still be dropped from the stack.Mark the genEnsure... stubs in CogObjRep as inlineand fix a signature to eliminate warnings.
VMMaker.oscog-eem.1302:Allow setting of debugBytecodePointers anddebugFixupPointers from initializationOptionsfor inimage compilation, allowing getting to thepoint quickly as inStackToRegisterMappingCogit	genAndDis: BorderedMorph>>#borderStyle	options: #(ObjectMemory NewCoObjectMemory				debugBytecodePointers #(137))Use the compiled method, not the header oop, whendecoding method headers during in-image compilation,so as not to be confused by the different compiledmethod header formats between V3 and Spur.

VMMaker.oscog-eem.1301:Cogit:Since the method zone does not scavenge PICs, theMNU method in an MNU PIC entry cannot be young.If it is, void the cache (load 0 instead of the method).