Cog-eem.309:Initial fixes for the Cuis Spur bootstrap.
Cog-eem.308:Clean up disassembly and use the decorateDisassembly:for:fromAddress: form on x64 to use relativeBaseForDisassemblyInto: for method-relative pc numbering.

Cog-eem.307:Do a better job of filtering uninitialized fp regs in x86 and x64 register state printing.

Cog-eem.306:Add all the fp registers to the x64 alien and provide a (paritally implemented) convenience to zero all the fp regs (for displaying regs each instruction, stripped of the noise of the fp regs).