travis: Modernize the travis file and gain building on OSX
travis-ci has gained OSX support and this requires a small
modernization of the file and approach. We can not just execute
sudo apt.. anymore but need to mention the packages we want and
they need to be whitelisted. All but the asound-plugins are in
the list and we are fine.

The OSX build is currently failing as the system doesn't have
the 10.6 SDK installed. This could be easily done by changing
the CMake VMMaker.