Cog-tpr.325:Correct the offset for the floating point CPSR
Cog-eem.324:Fix offsets for the fp regs in the ARM alien now that we compile the code generator with -DMOTET

Cog-eem.323:Complete the Spur bootstrap for Cuis 4.2. 3+4=7.Update some older prototypes to reflect the current state of code in Squeak 5.

Cog-tpr.322:Update ARM simulation of FP code; not complete as yet but mostly there.

Cog-eem.321:Add the missing wordSze method to the Cuis bootstrap and now we have a runner.