Allow to open a filedes and FILE* as SQFile*
When using BSD/Linux/Posix APIs a filedescriptor will be returned from
the system. Make it possible to turn either a filedes or a FILE* into a
SQFile*. This can be used to read/write using the StandardFileStream and
its subclasses.

A simple test to open filedes '2' and write to it has worked. More
advanced modes are still missing.