RegisterAllocatingCogit:
No longer flush on conditional branch; instead leave it up to the merge logic to harmonise the stacks on join.

Consequently, fix the order of generating fixups and pushing the dummyValue when no deadCode in genSpecialSelectorEqalsEquals:.  This didn't matter with the StackToRegisterMappingCogit but in the RegisterAllocatingCogit now that teh stack is not flushed on conditional branch, it is critical; the fixup must be established before the dummyValue is pushed.

Also consequently make mergeCurrentSimStackWith:/reconcileForwardsWith: more intelligent, able to merge when the target has spills and the current does not.

At least mark that mustBeBoolean: will switch to an interpreted frame (yet to be implementd).

Fix a slip in genSubConstant:R: which always used TempReg and didn't use its argument.

Fix a slip in temp name printing for sim stack printing.

Use = to compare objectmemory falseObject and objectMemory trueObject swith SSConstant entry constants.  These are integeger oops, not objects, and so integral comparison shoudl be used.
---
Name: VMMaker.oscog-eem.2028
Author: eem
Time: 6 December 2016, 6:08:48.407545 pm
UUID: a66fd6b5-4adb-4770-83b4-a29622cc4314
Ancestors: VMMaker.oscog-eem.2027