Cogit: Add support for SwapR:R:Scratch:, map it to XCHGRR on x64 & x86 and use it in the nascent parallel move code in RegisterAllocatingCogit. Fix a bug in x64's XCHGRR. Add generation tests for x64's XCHGRR.
Have printCogMethods print totals of the number of methods & PICs in the code zone.
Fix a slip in printCogMethodsWithPrimitive:.

RegisterAllocatingCogit:
Use ensureFixupAt: not ensureNonMergeFixupAt: for the targets of branches.  I was confused.  This gets the system much farther.
Nuke the check for no code being generated in ensureFixupAt:.  We're beyond that.
Deal with merging with a constant assigned to a register.
---
Name: VMMaker.oscogSPC-eem.2130
Author: eem
Time: 15 February 2017, 6:39:48.551135 pm
UUID: 646641b5-87bf-4dcd-85da-fdcff6123193
Ancestors: VMMaker.oscogSPC-eem.2129