unix: Remove copy and paste of mkNamedPrims.sh
For the debian package building I want/need to do

 mkdir build
 cd build
 create plugins.int/.ext
 ../platforms/unix/config/configure
 make

but since 4869c8d1fadffe9ed5a6b5c75cf3c59adc89fbe5 this does
not work and requires a very specific place and layout of the
build directory. Revert the commit and update the location of
the file. It will always be there in git.