platforms/unix/config/mkmf: rm without fully-qualified path
Replace '/bin/rm' with '\rm'. This makes the script work in any
environment that provides 'rm' in PATH and does not depend on the
specific path to the executable.

(Specifically this makes the script work on NixOS which supports
installing many versions of each program, such as rm, and uses PATH to
provide the right one.)

I chose '\rm' instead of 'rm' just in case the motivation for using
'/bin/rm' was to bypass a shell builtin. If that is not the case then
this could be simplified to just 'rm'. This wasn't clear to me from
the git history ('git blame').