unix configure: update configure in-place
I have manually updated the 'configure' script with the fix from commit
9072fbf0e188350c86a41205d0ef18de823d0a48 and added a loud warning that
the file is being maintained by hand.

I am assuming that somebody already decided to maintain this file by
hand? I tried running 'autoconf' with the exact same version as cited in
the file (2.59) but this ended up making a lot of changes compared with
the file in git.