Cogit: Use genMarshallNArgs:arg:arg:arg:arg: when compiling an interpreter primitive call to allow adding the shadow parameter space on Win64.
Fix generation of JumpR on x64.  I forgot a rex byte.  This doesn't affect SysV because we only ever jump through ClassReg (in an open PIC dispatch) which is RCX (1) on SysV but R8 (8) on Win64.

Add tests for CallR and JumpR generation on x64.

Have printRumpCStackTo: include the sp and fp if in the range.

With the fix to JumpR a Win64 image simulates correctly (with smashing of the shadow parameter space).
---
Name: VMMaker.oscog-eem.2225
Author: eem
Time: 29 May 2017, 9:34:28.198165 am
UUID: a3791777-370b-419b-9708-9e73bf94e39b
Ancestors: VMMaker.oscog-nice.2224