debian: Re-work packages and settle on new VM names
Re-work packages to have 32bit and 64bit x86 VMs to be installed
next to each other. This was the consensus between Esteban, Norbert
and me. Pick a solution that works for i386 and amd64. If we want
to support ARM (and AARCH64) we will need to have one more script
that will be executed at the install path that creates the final
structure (e.g. with $CC -print-multiarch) and remove i386/x86_64
from the .install files. This is left as an excercise for the reader.

pharo6-32 will be a 32bit VM for Pharo6 and maybe later
pharo6-64 will be a 64bit VM for Pharo6 and maybe later
pharo6-common-ui will install images and other files to be shared
pharo6-32-ui will install plugins required for the desktop
pharo6-64-ui will install plugins required for the desktop