travis-ci: Build and upload the debian package before the scripts
First I don't know how matrix and script within and outside interact
with each other but then just because it does or doesn't compile in
the CI doesn't say if it compiles for different versions of GCC. So
let's assume that when things hit master they generally compile. With
the before_script a failed source package upload will fail the build.

Clean things up after the upload so it looks like the script did not
run.